+++ src/components/atoms/InfoButton/cod-info-button.js
@@ -1,0 +1,2 @@
+import InfoButton from './InfoButton';
+customElements.define('cod-info-button', InfoButton);